Added CommentController, RatingController, CommentNotFoundAdvice, CommentNotFoundException, RatingNotFoundAdvice, RatingNotFoundException

+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/repositories/RatingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+public interface RatingRepository extends JpaRepository<Rating, Long> {
     Optional<Rating> findByGivenPoints(Integer givenPoints);
 }
